+++ client/src/components/Filters/index.js
@@ -1,0 +1,1 @@
+export { default as Filters } from './Filters';
