+++ client/src/components/Filters/styles/Filters.module.css
@@ -1,0 +1,3 @@
+.filters{
+
+}
